GridAR-EditAR/requirements.txt
torch==2.2.1
torchvision==0.17.1
torchaudio==2.2.1

# Qwen2.5-VL class support (4.44.2 is too old)
transformers @ git+https://github.com/huggingface/transformers@0ebd6651acd32c982fee265b23243b89bdb89577

# vllm==0.4.1
Pillow==10.0.1

# avoid NumPy 2 compatibility issues with some binary wheels
numpy==1.26.4

ftfy==6.2.3
datasets==2.21.0
accelerate==0.33.0
sentencepiece==0.2.0
pandas==2.2.2
gradio==4.42.0
bs4==0.0.2
einops==0.8.0
pycocotools==2.0.8
scikit-image==0.24.0
natsort==8.4.0
kornia==0.7.3
torchmetrics==1.5.1
torch-fidelity==0.3.0
diffusers==0.31.0
mmsegmentation>=1.0.0
clean-fid

# provides `from qwen_vl_utils import process_vision_info`
qwen-vl-utils==0.0.14
openai==1.107.1
protobuf==3.20.3